+++ lib/src/strings.dart
@@ -22,7 +22,6 @@ String get dylibFileName {
   return name;
 }
 
-const llvmLib = 'llvm-lib';
 const llvmPath = 'llvm-path';
 
 /// Name of the parent folder of dynamic library `lib` or `bin` (on windows).